rewrite fetch system to not rely on global overwrites
