+++ web/src/actions/Registration.ts
@@ -1,5 +1,7 @@
 'use server';
 
+import { serverFetch } from '../app/ServerUtils';
+
 export async function register(
     email: string,
     username: string,
@@ -9,7 +11,7 @@ export async function register(
         status: number;
     } & ({ ok: false; message: string } | { ok: true })
 > {
-    const res = await fetch('/api/registration/register', {
+    const res = await serverFetch('/api/registration/register', {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json',
