+++ web/src/app/(main)/staff/@tokens/CreateTokenForm.tsx
@@ -1,28 +1,34 @@
 'use client';
-import { useState } from 'react';
 import { createToken } from '../../../../actions/ApiTokens';
-import { useRouter } from 'next/navigation';
+import { Form, Formik } from 'formik';
+import { Button } from '@mui/material';
+import { object, string } from 'yup';
+import FormikTextField from '../../../../components/input/FormikTextField';
+
+const tokenValidationSchema = object().shape({
+    name: string().required('Application name is required'),
+});
 
 export default function CreateTokenForm() {
-    const [name, setName] = useState('');
-    const router = useRouter();
-
     return (
-        <>
-            <input
-                type="text"
-                name="name"
-                onChange={(e) => setName(e.target.value)}
-                value={name}
-            />
-            <button
-                onClick={() => {
-                    createToken(name);
-                    router.refresh();
-                }}
-            >
-                Create Token
-            </button>
-        </>
+        <Formik
+            initialValues={{
+                name: '',
+            }}
+            validationSchema={tokenValidationSchema}
+            onSubmit={async ({ name }) => {
+                const res = await createToken(name);
+            }}
+        >
+            <Form>
+                <FormikTextField
+                    id="token-name"
+                    type="text"
+                    name="name"
+                    label="Name"
+                />
+                <Button>Create Token</Button>
+            </Form>
+        </Formik>
     );
 }
